SpringBoot_Project/src/test/java/com/mayur/Java_crud/service/ProductServiceTest.java
package com.mayur.Java_crud.service;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.Mockito.when;
import static org.mockito.Mockito.times;
import static org.mockito.Mockito.verify;


import java.util.Arrays;
import java.util.List;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.MockitoAnnotations;

import com.mayur.Javacrud.model.Product;
import com.mayur.Javacrud.repository.ProductRepository;
import com.mayur.Javacrud.service.ProductService;

public class ProductServiceTest {

	@Mock
	private ProductRepository productRepository;

	@InjectMocks
	private ProductService productService;

	@BeforeEach
	public void setUp() {
	MockitoAnnotations.openMocks(this);
	}
	
	@Test
    public void testGetAllProducts() {
        List<Product> productList = Arrays.asList(
                new Product(1L, "Product 1", 10.0),
                new Product(2L, "Product 2", 20.0)
        );

        when(productRepository.findAll()).thenReturn(productList);

        List<Product> result = productService.getAllProducts();

        assertEquals(2, result.size());
        verify(productRepository, times(1)).findAll();
	}
	// Can Write more test cases for other methods
}
